Simplify buildTools tests: drop Tool/JVMSupport abstraction
Collapse the single-tool Gradle suite hierarchy into one concrete class
and replace the dead Tool/JVMSupport matrix with a maxJdk constraint.
